remove envelopestatus metadata.
Now that envelopestatus is a Datum, ctime and mtime (especially ctime)
probably can be used for 'received' instead of holding it
explicitly. We also get UUID for free since, again, it's a Datum

This may be a useful place to put tags or starring. The reason for
seperating this from "Envelope" is because things like starring and
tagging are less likely to be want to be pushed to other users,
whereas the Envelope will be passed verbatim (wrapped or otherwise) in
the common case.

Signed-off-by: Dan Farina <drfarina@acm.org>